requirements.txt
asgiref==3.3.4
Django==3.2.3
django-redis==4.12.1
djangorestframework==3.12.4
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.2.0
